+++ pwa/src/pages/Feed/Feed.js
@@ -1,9 +1,8 @@
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import styled from "styled-components";
-import * as S from "../Word/Word.style.js";
-import variables from "../../styles/variables";
 import useStore from "../../state/store";
+import * as F from "./Feed.style.js";
+import * as S from "../Word/Word.style.js";
 
 const Feed = () => {
   const [myFeed, setMyFeed] = useState([]);
@@ -62,60 +61,60 @@ const Feed = () => {
 
   return (
     <>
-      <FeedPageHeader
+      <F.FeedPageHeader
         style={{ backgroundImage: getBackgroundImage(currentMemory) }}
       >
-        <UserInfo>내 정보</UserInfo>
-        <FeedSection>
-          <MyMemorySection value={currentMemory} onClick={MyMemory}>
+        <F.UserInfo>내 정보</F.UserInfo>
+        <F.FeedSection>
+          <F.MyMemorySection value={currentMemory} onClick={MyMemory}>
             나의 추억
-          </MyMemorySection>
-          <FriendsMemorySection value={currentMemory} onClick={FriendsMemory}>
+          </F.MyMemorySection>
+          <F.FriendsMemorySection value={currentMemory} onClick={FriendsMemory}>
             친구의 추억
-          </FriendsMemorySection>
-        </FeedSection>
-      </FeedPageHeader>
-      <FeedFrameContainer>
+          </F.FriendsMemorySection>
+        </F.FeedSection>
+      </F.FeedPageHeader>
+      <F.FeedFrameContainer>
         {(currentMemory === 0 ? myFeed : friendsFeed).map((data) => {
           const sliceData = (a, b) => {
             return data.createdAt?.slice(a, b);
           };
           return (
-            <FeedFrame key={data.id}>
-              <FeedInfo>
-                <FeedOwnerInfo>
-                  <FeedOwnerImg src={data.profileImageUrl} />
-                  <FeedOwnerName>{data.author}</FeedOwnerName>
-                </FeedOwnerInfo>
-                <FeedUploadDate>
+            <F.FeedFrame key={data.id}>
+              <F.FeedInfo>
+                <F.FeedOwnerInfo>
+                  <F.FeedOwnerImg src={data.profileImageUrl} />
+                  <F.FeedOwnerName>{data.author}</F.FeedOwnerName>
+                </F.FeedOwnerInfo>
+                <F.FeedUploadDate>
                   {`${sliceData(0, 4)}년 ${sliceData(5, 7)}월 ${sliceData(
                     8,
                     10
                   )}일`}
-                </FeedUploadDate>
-              </FeedInfo>
-              <FeedImg src={data.pictureDiary} alt="새록새록" />
+                </F.FeedUploadDate>
+              </F.FeedInfo>
+              <F.FeedImg src={data.pictureDiary} alt="새록새록" />
 
-              <FeedContent>
-                <FeedWord>{data.keyWord}</FeedWord>
-                <FeedDetailButton
+              <F.FeedContent>
+                <F.FeedWord>{data.keyWord}</F.FeedWord>
+                <F.FeedDetailButton
                   onClick={() => {
                     navigate("/feedDetail");
                     feedDetailData.push(data);
                   }}
                 >
                   자세히 보기 {">"}
-                </FeedDetailButton>
-              </FeedContent>
-              <FeedText>
+                </F.FeedDetailButton>
+              </F.FeedContent>
+              <F.FeedText>
                 {data.textDiary.length > 38
                   ? data.textDiary.slice(0, 37) + "..."
                   : data.textDiary}
-              </FeedText>
-            </FeedFrame>
+              </F.FeedText>
+            </F.FeedFrame>
           );
         })}
-      </FeedFrameContainer>
+      </F.FeedFrameContainer>
       <S.NavBar>
         <S.NavButton onClick={() => navigate("/word")}>
           <S.NavButtonIcon
@@ -137,131 +136,3 @@ const Feed = () => {
 };
 
 export default Feed;
-
-const FeedPageHeader = styled.div`
-  ${variables.position("fixed", 0, "null", "null", 0)};
-  ${variables.widthHeight("100%", "187px")};
-  display: grid;
-
-  background-size: 100% 100%;
-  background-repeat: no-repeat;
-
-  @media (min-width: 769px) {
-    ${variables.widthHeight("375px", "187px")};
-    ${variables.position("sticky", 0, "null", "null", 0)};
-    margin: calc(-20px);
-    transform: translate(0%, -11%);
-  }
-`;
-
-const UserInfo = styled.div`
-  ${variables.fontStyle("19px", 500)};
-  position: relative;
-  padding: 61px 19px 0 0;
-  color: ${(props) => props.theme.style.gray5};
-  text-align: right;
-  line-height: 29px;
-  letter-spacing: -0.03em;
-  cursor: pointer;
-`;
-
-const FeedSection = styled.div`
-  ${variables.flex("row", "null", "baseline")}
-  ${variables.widthHeight("100%", "57px")};
-`;
-
-const MyMemorySection = styled.div`
-  ${variables.fontStyle("19px", 600)};
-  width: 50%;
-  color: ${(props) => (props.value === 0 ? "#2f2f2f" : " #7d7d7d")};
-  text-align: center;
-  letter-spacing: -0.03em;
-  cursor: pointer;
-`;
-
-const FriendsMemorySection = styled.div`
-  ${variables.fontStyle("19px", 600)};
-  width: 50%;
-  color: ${(props) => (props.value === 1 ? "#2f2f2f" : " #7d7d7d")};
-  line-height: 29px;
-  text-align: center;
-  letter-spacing: -0.03em;
-  cursor: pointer;
-`;
-
-const FeedFrameContainer = styled.div`
-  margin: 169px calc(-20px) 168px;
-
-  @media (min-width: 769px) {
-    margin-top: 22px;
-  }
-`;
-
-const FeedFrame = styled.div`
-  ${variables.widthHeight("100%", "430px")}
-  margin : 0  auto 42px;
-  overflow: hidden;
-  display: grid;
-  justify-items: center;
-`;
-
-const FeedInfo = styled.div`
-  ${variables.flex("row", "space-between", "center")}
-  width : 335px;
-`;
-
-const FeedOwnerInfo = styled.div`
-  ${variables.flex()}
-  gap : 8px;
-`;
-
-const FeedOwnerImg = styled.img`
-  ${variables.widthHeight("32px", "32px")}
-  background-color: ${(props) => props.theme.style.gray1};
-  border-radius: 50%;
-`;
-
-const FeedOwnerName = styled.div`
-  ${variables.fontStyle("19px", 600)};
-  line-height: 29px;
-  letter-spacing: -0.03em;
-  color: ${(props) => props.theme.style.black};
-`;
-
-const FeedUploadDate = styled.div`
-  ${variables.fontStyle("19px", 500)};
-  line-height: 29px;
-  text-align: right;
-  letter-spacing: -0.03em;
-  color: ${(props) => props.theme.style.gray3};
-`;
-
-const FeedImg = styled.img`
-  ${variables.widthHeight("335px", "242px")}
-  margin-top : 20px;
-  background-color: ${(props) => props.theme.style.gray1};
-`;
-
-const FeedContent = styled.div`
-  ${variables.flex("row", "space-between", "center")};
-  width: 335px;
-  margin: 24px 0 16px 0;
-`;
-
-const FeedWord = styled.div`
-  ${variables.fontStyle("24px", 500)};
-`;
-
-const FeedDetailButton = styled.div`
-  ${variables.fontStyle("19px", 500)};
-  color: #828282;
-  cursor: pointer;
-`;
-
-const FeedText = styled.div`
-  ${variables.widthHeight("335px", "62px")}
-  ${variables.fontStyle("22px", 500)};
-  line-height: 32px;
-  letter-spacing: -0.03em;
-  color: ${(props) => props.theme.style.gray5};
-`;
